move get_latest_blocknumber to index_aggregate_user
